requirements.txt
aioredis==1.2.0
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.3
channels==2.4.0
channels-redis==2.3.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.38.0
django-cors-headers==3.5.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-taggit==1.3.0
djangorestframework==3.12.1
hiredis==0.2.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.5.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.5.6
mysqlclient @ file:///C:/Users/ANDALAN/Desktop/Python%20Package/mysqlclient-1.4.6-cp38-cp38-win32.whl
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==19.1.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.11
wrapt==1.12.1
zope.interface==5.1.2